pyproject.toml
[tool.poetry]
name = "locker"
version = "0.1.0"
description = "Project based on Djangop Connected with IOT Devices to Open and Close lockers"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.10.12"
djangorestframework = "^3.14.0"
Django = "^4.2.5"
psycopg2-binary = "^2.9.7"
razorpay = "^1.4.1"
python-dotenv = "^1.0.0"
drf-yasg = "^1.21.7"
paho-mqtt = "^1.6.1"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"

[tool.poetry.dev-dependencies]
httpie = "^3.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
